deploy_k8_fastapi/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
pillow==10.0.1
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0